Continued to Make Progress on the "Allow Signed in Users to Create Posts" Feature (Ticket #1)
Ticket #1: Only logged in users are now able to Create Posts which will be added to the homepage. When the Post button is clicked on the homepage, users are redirected to the Post-Create Form where users (logged in only) can create posts by entering values for the title, content/description, location, quantity, and price (if the user is not logged in, then they are redirected to the sign-in page). Also, if any user clicks on the title of a Post, then they will be redirected to the Post Detailed View Page.