Completed the "Allow Signed in Users to Create Posts" Features (Ticket #1)
Ticket #1: Made it so that pictures can be added to the Posts created by users. In addition, these pictures are optional (no user is required to upload a picture when creating a post). The Posts that contain a picture will now have that picture shown on the homepage with its associated post.