Completed the "Search By Title or Location" Features (Ticket #5)
Ticket #5: Made it so that users (logged in or not) can Search for Posts by their “Title” or “Location”. All of the posts with the matching “Title” or “Location” are then displayed. If no posts were found by the Search, the message “No posts were found. Please search for something else!” is displayed.