Completed the “Allow Signed in Users to Bookmark Posts” and “View all Bookmarked Posts” Features (Ticket #1 and #2)
Ticked #1 and #2: The database has been updated so that each post can be bookmarked. The ability for signed in users to bookmark posts created by other users has been created. These bookmarked posts can then be viewed on a separate tab titled “Bookmarked Posts” which is accessible to signed in users only (as only they have the ability to bookmark posts). More specifically, it has been made so that posts in the “Bookmarked Posts” tab are visible in a grid view with two columns. In addition, the post detailed view has been updated to include an “Add Bookmark” button. Note that this button is only visible to signed in users who are not the author of the post or have already added this post to their list of bookmarked posts. Moreover, logged in users who have no bookmarked posts will see the message “You Have no Bookmarked Posts!” if they click on the “Bookmarked Posts” tab (logged out users will not have the “Bookmarked Posts” tab visible). Furthermore, the CSS on the Post-Create Form has been updated so that it looks cleaner. Also, made it so that if the user clicks on the title of a Post within the “Bookmarked Posts” tab, then they will be redirected to the Post Detailed View Page. Additionally, the following tests have been done: Tested to see that this button no longer appears if the post has already been bookmarked by the user. Tested so that users who are logged out do not see the “Add Bookmark” button. Tested so that if the user is the owner of the post, then they will not see the “Add Bookmark” button, and thus, are not able to bookmark their own posts. Tested so that if the user logs out and back in, the bookmarked posts are still there in the “Bookmarked Posts” tab.